Refactor: moved api definition from cmd/insprd to pkg
