+++ cmd/insprd/auth/payload.go
@@ -1,6 +1,6 @@
 package auth
 
-//Payload is information caried by a Inspr acceess token
+//Payload is information caried by a Inspr access token
 type Payload struct {
 	UID        string
 	Role       int
