+++ pkg/api/controllers/routes.go
@@ -3,7 +3,7 @@ package controller
 import (
 	"net/http"
 
-	handler "gitlab.inspr.dev/inspr/core/cmd/insprd/api/handlers"
+	handler "gitlab.inspr.dev/inspr/core/pkg/api/handlers"
 )
 
 func (s *Server) initRoutes() {
